[BE] refactor: group 레포 리팩토링
- group 단 레포 리팩토링
